Add support for wide characters
Currently, the width of text strings are assumed to be the same as their
byte length. In effect, this means that only ASCII is displayed at its
correct width.

This commit fixes the issue by explicitly determining the display width
of each string or character when calculating widget size and drawing
text to the screen.