Fix Painter to have a correct style stack.
Without this commit, WithStyle makes a Painter reset to the default
rather than going back to the style it had before. This means if a
a Widget within a styled Box is also styled, the Widgets after it in the
Box would not inherit the Box's style.

Previous commits demonstrate this bug; this gets the tests to pass.