Move most _test files into a tui_test package.
There's a few articles positing this is best-practice, as it encourages
a focus on external behavior and APIs rather than implementation
details. I tend to agree.

A few tests are based on internal-only behaviors, or types which are public
but are effectively internal-only. These were kept in the `tui` package,
with _internal_test.go suffixes.

I did have to tweak a couple tests to get this split to work - making
entry.Offset public and skipping some of the other positioning tests.