Make Style's stacking behavior explicit.
The previous Style behavior was for each WithStyle application to reset
each property; application of "this color" and then "underline" would go
back to ColorDefault.

This allows each of Style's properties to be either explicitly
specified, or to inherit from the existing Style. The default is to
inherit from the next style up in the widget stack.

Mostly this is accomplished by making the boolean properties into a
tri-state (inherit/on/off) type.