Demonstrate "default style" != "zero Style".
WithStyle and Theme have an interesting behavior. Theme.Style will
always return a value (the zero value of Style, essentially), even if
the Style with the requested name doesn't have anything set in the
Theme. However, as far as I can tell, this does't matter- WithStyle
first checks if the Theme HasStyle, and only applies an additional style
when one is set in the Theme.

I'd like these to be the same (see marcusolsson/tui-go#81)... but I'm
not going to send this as a PR yet until that's agreed.