Factor wrappedText and fix tests.
The `SizeHint` method was giving incorrect results because it was
unconditionally giving the hint for wrapped text, via calling
`heightForWidth` even when `l.wordWrap` wasn't set. This was
particularly causing Unicode tests to fail.

So: replaced `heightForWidth` with the more useful `lines`
accessor, which automatically gets the wrapped-or-not text of the
appropriate size, split into lines. That's what `Draw` and `SizeHint`
both need.