Update docstring on Padder
I was confused as to the semantics of `x` and `y`: does it pad *to* the specified size, e.g. "be sure you're at least X by Y", or is it "this much space around the outside"?

Apparently it's the latter, i.e. CSS style (without being able to specify top-and-bottom).