Make Update synchronous.
Before this change, Update is a maybe-blocking operation. Specifically,
if there's an Update queued already, your Update is going to block;
however, if the queue is empty, it will return immediately... without
actually running the function.

The worst case is if Update is called within the event thread itself. In
that case, Update may be able to be called exactly once... and then the
second time, it will deadlock.

This change
- Makes the callbackFn channel unbuffered
- Wraps callbacks with a blocking channel

in order to make Update a synchronous operation. If a user wants to run
Update asynchronously, that's easy - just spell it `go ui.Update`
instead.

This solution brought to you by Slack discussion with @marcusolsson and
@ingcr3at1on, and also by
http://journal.stuffwithstuff.com/2015/02/01/what-color-is-your-function/.