Add a test demonstrating that 128 is magic.
According to this test, when there are nested HBoxen, they don't get
as wide as they could - tops out at around 128 characters. (Not a
coincidence.)

This is weird and sad, because a full-width terminal can certainly be
>128 characters.

I suspect the same would be true of tall displays as well, but don't
have one sufficiently tall handy.